requirements.txt
asgiref==3.2.10
Babel==2.8.0
Django==3.0.8
django-phonenumber-field==4.0.0
django-widget-tweaks==1.4.8
phonenumbers==8.12.7
pytz==2020.1
sqlparse==0.3.1
